+++ sparse_autoencoder/optimizer/abstract_optimizer.py
@@ -2,7 +2,7 @@
 from abc import ABC, abstractmethod
 from typing import TypeAlias
 
-from jaxtyping import Int
+from jaxtyping import Int64
 from torch import Tensor
 from torch.nn.parameter import Parameter
 from torch.optim import Optimizer
@@ -33,7 +33,7 @@ class AbstractOptimizerWithReset(Optimizer, ABC):
     def reset_neurons_state(
         self,
         parameter: Parameter,
-        neuron_indices: Int[Tensor, Axis.LEARNT_FEATURE_IDX],
+        neuron_indices: Int64[Tensor, Axis.LEARNT_FEATURE_IDX],
         axis: int,
     ) -> None:
         """Reset the state for specific neurons, on a specific parameter.
